// apps/studio/src/lib/live/liveParams.gen.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface LiveParams {
  filter_cutoff?: number;
  filter_res?: number;
  drive?: number;
  delay_time?: number;
  echo_feedback?: number;
  reverb_mix?: number;
  output_level?: number;
  fx_preset?: "clean" | "dub" | "big_room" | "radio" | null;
  ts: number;
}